requirements.txt
ptpython
pytest
pillow
numpy
matplotlib